Move EventArgs class to new files.
